dental_counseling_ai/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
pandas==2.1.3
openai==1.3.7
python-docx==1.1.0
pydub==0.25.1
SpeechRecognition==3.10.0